Mostly done

+++ app/frontend/vue/news/news-app.vue
@@ -1,5 +1,8 @@
 <template>
-  <div>
+  <b-container>
+    <b-button class="mb-4" variant="outline-primary" @click="isDialogOpen = true">
+      Neue News
+    </b-button>
     <b-table :items="news" :fields="fields">
       <template #cell(message)="d">
         <span class="abbreviate">{{ d.item.message }}</span>
@@ -8,14 +11,14 @@
         <div class="text-nowrap">
           <b-button
             variant="outline-success"
-            class="edit-user btn btn-sm mr-1"
+            class="edit-user btn btn-sm me-1"
             @click="showNews(data.item)">
             <i class="fas fa-edit"></i>
             <span class="d-none d-lg-inline">Anzeigen</span>
           </b-button>
           <b-button
             variant="outline-danger"
-            class="edit-user btn btn-sm mr-1"
+            class="edit-user btn btn-sm me-1"
             @click="deleteNews(data.item)">
             <i class="fas fa-trash"></i>
             <span class="d-none d-lg-inline">Löschen</span>
@@ -23,13 +26,43 @@
         </div>
       </template>
     </b-table>
-  </div>
+    <confirm-dialog ref="confDialog" />
+    <b-modal
+      ref="createNewsDialog"
+      v-model="isDialogOpen"
+      title="News erstellen/bearbeiten"
+      @hidden="resetDialog">
+      <b-form>
+        <b-form-input
+          v-model="title"
+          placeholder="Titel eingeben (optional)"
+          class="mb-2"></b-form-input>
+        <b-form-textarea
+          v-model="message"
+          placeholder="Nachricht eingeben"
+          class="mb-2"></b-form-textarea>
+        <b-button variant="outline-danger" class="me-2" @click="isDialogOpen = false">
+          Abbrechen
+        </b-button>
+        <b-button variant="outline-success" @click="saveNews">Speichern</b-button>
+      </b-form>
+    </b-modal>
+  </b-container>
 </template>
 <script setup>
   import { ajax } from 'src/utils/ajax'
   import { onBeforeMount, ref } from 'vue'
+  import ConfirmDialog from 'src/vue/shared/confirm-dialog.vue'
 
   const news = ref()
+  const selectedNews = ref()
+
+  const title = ref('')
+  const message = ref('')
+  const isDialogOpen = ref(false)
+
+  const confDialog = ref(0)
+  const createNewsDialog = ref(0)
 
   const fields = [
     { key: 'id', label: 'Id' },
@@ -40,15 +73,53 @@
   ]
 
   onBeforeMount(async () => {
-    const res = await ajax({ url: '/news' })
-    news.value = res.data
+    fetch()
   })
 
-  const showNews = id => {}
+  const fetch = async () => {
+    const res = await ajax({ url: '/news' })
+    news.value = res.data
+  }
+
+  const showNews = item => {
+    selectedNews.value = news.value.find(n => n.id === item.id)
+    title.value = selectedNews.value.title
+    message.value = selectedNews.value.message
+    isDialogOpen.value = true
+  }
+
+  const saveNews = async () => {
+    console.log('selected', selectedNews.value)
+    if (!selectedNews.value) {
+      await ajax({
+        url: '/news',
+        method: 'POST',
+        data: { news: { title: title.value, message: message.value } },
+      })
+    } else {
+      await ajax({
+        url: `/news/${selectedNews.value.id}`,
+        method: 'PATCH',
+        data: { news: { title: title.value, message: message.value } },
+      })
+    }
+    await fetch()
+  }
 
   const deleteNews = async item => {
-    console.log('items', item)
-
-    await ajax({ url: `/news/${item.id}`, method: 'DELETE' })
+    const ok = await confDialog.value.open({
+      title: 'Messung löschen',
+      message: `Diese Messung unwiderruflich löschen! Sind Sie sicher?`,
+      okText: 'Ja, löschen',
+    })
+    if (ok) {
+      const res = await ajax({ url: `/news/${item.id}`, method: 'DELETE' })
+      news.value = res.data
+    }
+  }
+  const resetDialog = () => {
+    selectedNews.value = undefined
+    title.value = ''
+    message.value = ''
   }
 </script>
